# src/main/resources/static/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "TGM Schulplaner API",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/v1/user": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get the authenticated user",
        "operationId": "getUserInfo",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer-key": [

            ]
          }
        ]
      }
    },
    "/api/v1/users": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get all users filtered by q and limited by size and page if the user is authenticated",
        "operationId": "searchForUser",
        "parameters": [
          {
            "name": "q",
            "in": "query"
          },
          {
            "name": "page",
            "in": "query"
          },
          {
            "name": "size",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDTO"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer-key": [

            ]
          }
        ]
      }
    },
    "/api/v1/group/{gid}/users": {
      "get": {
        "tags": [
          "group",
          "user"
        ],
        "summary": "Get all users that are a member of the group if the authenticated user has access",
        "operationId": "getGroupUsers",
        "parameters": [
          {
            "name": "gid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDTO"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer-key": [

            ]
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "UserDTO": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "department": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "bearer-key": {
        "type": "http",
        "name": "bearer-key",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "jwt"
      }
    }
  }
}